Do not use CPPFLAGS to add headers to source files
Source code, e.g. #include, belongs in the source code, not the build
system. Build systems are for defining include paths, not the includes
themselves.

Drop the KeySym keyword since it seems to be a useless alias for
uint32_t which might as well be used directly instead of indirectly
using #define everywhere.